fix minor race issue when waiting for condition
scheduler sets nodemaintenance to scheduled state in condition
then waits for it to be updated in cache.

between polling intervals a nodemaintenance may already be transitioned
to a more advance state e.g draining or ready.

instead, check if node is under maintenance to match on any other state
than uninitialized or pending.

Signed-off-by: adrianc <adrianc@nvidia.com>